Move Stick class into a globally available library
