+++ src/modules/flight_mode_manager/tasks/Descend/FlightTaskDescend.hpp
@@ -37,9 +37,9 @@
 
 #pragma once
 
+#include <lib/sticks/Sticks.hpp>
 #include <lib/stick_yaw/StickYaw.hpp>
 #include "FlightTask.hpp"
-#include "Sticks.hpp"
 #include "StickTiltXY.hpp"
 
 class FlightTaskDescend : public FlightTask
