+++ src/modules/flight_mode_manager/tasks/ManualAltitude/FlightTaskManualAltitude.hpp
@@ -40,8 +40,8 @@
 #pragma once
 
 #include <lib/stick_yaw/StickYaw.hpp>
+#include <lib/sticks/Sticks.hpp>
 #include "FlightTask.hpp"
-#include "Sticks.hpp"
 #include "StickTiltXY.hpp"
 #include <uORB/Subscription.hpp>
