+++ src/modules/flight_mode_manager/tasks/Utility/CMakeLists.txt
@@ -32,13 +32,12 @@
 ############################################################################
 
 px4_add_library(FlightTaskUtility
-	Sticks.cpp
 	StickAccelerationXY.cpp
 	StickTiltXY.cpp
 	Gimbal.cpp
 )
 
-target_link_libraries(FlightTaskUtility PUBLIC FlightTask hysteresis SlewRate motion_planning mathlib)
+target_link_libraries(FlightTaskUtility PUBLIC FlightTask hysteresis SlewRate Sticks motion_planning mathlib)
 target_include_directories(FlightTaskUtility PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
 
 px4_add_functional_gtest(SRC StickTiltXYTest.cpp LINKLIBS FlightTaskUtility)
