+++ src/modules/flight_mode_manager/tasks/Utility/Gimbal.hpp
@@ -48,7 +48,8 @@
 #include <uORB/topics/gimbal_manager_status.h>
 #include <uORB/topics/vehicle_command.h>
 
-#include "Sticks.hpp"
+#include <lib/sticks/Sticks.hpp>
+
 
 
 class Gimbal : public ModuleParams
